More changes
(1) rename and add constructor functions
(2) use EnumSet for the character classes

Initially I used the suffix "All" to indicate greedy matching,
but the word "all" is so generic it doesn't really add much to
readability when in this case it means something entirely
different and specific. Using other words equally feels off,
misleading or verbose.
If I have to presuppose familiarity of regex terms,
adding "?" perfectly conveys that the match is lazy.
Or not, I'm just trying to rationalize being able to
use special characters in an identifier, because
why not.

(2) I opted to use EnumSet since I need to combine the character
classes later on when normalizing.