Use immutable linked list for captures
I initially used ArrayListUnmanaged for it,
but I thought I could avoid heap allocations when
capturing matches using just linked list.
Probably a terrible idea, more than a curious
exercise actually. But I liked how it turned out,
no cross-function mutations, so it's (feels) easier to reason about.

It was tempting to do the same with * or + matching,
but unlike captures where there are expectedly small
fixed number of items, repetitions are variable and
could easily grow in the size of the input string,
so I'd get stackoverflow and yo mama jokes if I do that.