Invoke @panic on allocation error
As stated in the added comment, allocation
errors are no longer explictly handled,
and instead panic is invoked. This means
no more !, try or errdefer in sight,
and the code is now simpler.

I'm not implying that explicity error handling
is bad. But in this case, the only
error that can occur is when heap allocating,
and failure to allocate due to insufficient
memory is unrecoverable,
so there's no reason to bubble up the
error from caller to caller.