Improve and fix captures
Previously captures are only avaiable internally
for back-references. I avoided making any heap allocations
during match()'ing, and that includes the captures.
I managed to keep a list of captures in the stack memory
without pre-allocating by using the call stack,
which is to say, I call a new function for every
capture made:

            MatchState.Capture

(stack)
capture <-|
  .prev   |                          (stack frame)
          |
----------+---------------------------------------
          |  (stack)
          |  capture <-|
          ---- .prev   |             (stack frame)
                       |
-----------------------+--------------------------
                       |  (stack)
                       |  capture
                       ---- .prev    (stack frame)

To make captures available for the
the caller, they either needed to be passed using
a closure, or with heap allocation. I'm planning
on adding threads later on to parallelize searching,
and I thought closures probably wouldn't fit that
model well, so opted to use heap allocations.

Since the resulting captures will be done in a
supposed multi-threaded environment, I can't
mutate the returned heap allocated captures directly.
After some head-scratching and false starts, I went
with the following:

            MatchResult.Capture
(heap)
capture
  .next --|                           (stack frame)
          |
----------+----------------------------------------
          |   (heap)
          |-> capture
                .next --|             (stack frame)
		        |
------------------------+--------------------------
                        |   (heap)
                        |-> capture
                              .next   (stack frame)

I used to separate types of captures to convey the
intention of different usages.

I also used ArenaAllocator for the result
captures since it would be easier to discard
temporary allocations that way.