Remove arena allocator for captures
Previously I used arena for allocating temporary captures,
as I needed to copy and avoid modifying them as other
threads may be referring to them too. But after some
reconsideration, this isn't really needed since
the caller has ownership of the captures. This is because
the functions that spawn threads also waits for them to
finish before returning.

Aside from that, make some changes so that the user of the API
don't have to manually free the captures with the iterator.

Also add skipped method too for more fine-grained search-and-replace
use cases.